Put lastFire table config in data.js
